ddms\classes\command\NewDynamicOutputComponent: Continued development of ddms\class\command\NewDynamicOutputComponent. Implemented test, testRunDoesNotCreateDynamicOutputFileInAppsDynamicOutputDirectoryIfDynamicOutputFileAlreadyExists(). All PhpUnit and PhpStan tests are passing. This is related to addressing issue #12.

+++ ddms/classes/command/NewDynamicOutputComponent.php
@@ -26,13 +26,17 @@ class NewDynamicOutputComponent extends AbstractCommand implements Command
             $this->pathToNewDynamicOutputComponent($flags),
             $this->generateDynamicOutputComponentFileContent($flags)
         );
-        if(isset($flags['shared']) && !file_exists($this->determineSharedDynamicOutputDirectoryPath($flags))) {
-           mkdir($this->determineSharedDynamicOutputDirectoryPath($flags));
+
+        if(!file_exists($this->pathToNewDynamicOutputFile($flags))) {
+
+            if(isset($flags['shared']) && !file_exists($this->determineSharedDynamicOutputDirectoryPath($flags))) {
+               mkdir($this->determineSharedDynamicOutputDirectoryPath($flags));
+            }
+            file_put_contents(
+                $this->pathToNewDynamicOutputFile($flags),
+                ''
+            );
         }
-        file_put_contents(
-            $this->pathToNewDynamicOutputFile($flags),
-            ''
-        );
     }
 
     /**
